pyproject.toml
[tool.poetry]
name = "home work 02-12"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "home work 02"}]

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = ".git"

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = ".venv"
